refactor(infrastructure): PR-7 commit 2 — secrets.py 763 搬 infrastructure/
简化版：仅搬迁，不做 3-way 拆（models/store/migrations）。

**planning 决策偏离**：原 planning 写 "secrets.py 拆 infrastructure/secrets/
{models,store,migrations}.py（170 行 legacy migration 保留但隔离）"，本 PR
**只搬迁不拆**。原因：

- Pydantic v2 validator 跨文件循环风险高 —— Secrets BaseModel 引 LLMTaggerConfig
  / WD14Config 等近 20 个 model，且 `_migrate_legacy_schema` 170 行涉及大量
  字段名穿插，3-way 切分需要前向引用 + 类型 hint 字符串化才能跨文件 work
- 现状单文件 763 行不影响读 / 不影响日常修改；split 收益主要是"心理"层
  （legacy migration 视觉隔离），延后无碍
- PR-7 整体范围已经够大；3-way 拆推到独立 PR 更稳

  studio/infrastructure/secrets.py  763 行整搬迁（git rename detection 100%）
  studio/secrets.py                 11 行 shim（sys.modules 别名）

回归：PR-1 安全网 152 + secrets 2 + studio_server 44 + ... = **285/285 全绿**。
